fix bug

+++ fontend/web/Blog-Dependency/src/app/modules/posts/post-add/post-add.component.ts
@@ -2,7 +2,10 @@ import { Component, Input, OnChanges, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { faMarkdown } from '@fortawesome/free-brands-svg-icons';
+import { AccountLogin } from 'src/app/model/acccount-login';
+import { Account } from 'src/app/model/account';
 import { Post } from 'src/app/model/post';
+import { AccountService } from 'src/app/service/account.service';
 import { DateService } from 'src/app/service/date.service';
 import { PostService } from 'src/app/service/post.service';
 
@@ -17,15 +20,20 @@ export class PostAddComponent implements OnInit, OnChanges {
   faMarkdown = faMarkdown;
   formPost: FormGroup;
   postPreview = '';
+  currentAccount: AccountLogin;
+
   constructor(
     private fb: FormBuilder,
     private postService: PostService,
     private dateService: DateService,
     private route: ActivatedRoute,
-    private router: Router
-  ) { }
+    private router: Router,
+    private accountService: AccountService
+  ) {
+  }
 
   ngOnChanges(): void {
+    this.currentAccount = this.accountService.getCurrentAccountValue;
     this.formPost = this.fb.group({
       heading: [this.flagEdit ? this.postEditting.heading : '', [Validators.required]],
       subHeading: [this.flagEdit ? this.postEditting.subHeading : '', [Validators.required]],
@@ -36,15 +44,21 @@ export class PostAddComponent implements OnInit, OnChanges {
     this.postPreview = this.formPost.get('content').value;
   }
   ngOnInit(): void {
+    this.currentAccount = this.accountService.getCurrentAccountValue;
 
   }
 
   submitPost(): void {
+    if (!this.currentAccount) {
+      return;
+    }
     const post = new Post();
     post.heading = this.formPost.get('heading').value;
     post.subHeading = this.formPost.get('subHeading').value;
     post.content = this.formPost.get('content').value;
     post.urlImage = this.formPost.get('urlImage').value;
+    post.accountDTO = this.postEditting.accountDTO;
+
     if (this.flagEdit) {
       post.id = this.postEditting.id;
       post.dateCreate = this.postEditting.dateCreate;
@@ -78,6 +92,9 @@ export class PostAddComponent implements OnInit, OnChanges {
     }
   }
   deletePost(): void {
+    if (!this.currentAccount) {
+      return;
+    }
     const idDelete: number = this.postEditting.id;
     this.postService.deletePost(idDelete).subscribe(post => {
       console.log(post);
